+++ src/libtrx/game/viewport.c
@@ -1,8 +1,12 @@
 #include "game/viewport.h"
 
 #include "config.h"
+#include "game/math/const.h"
+#include "game/output/vars.h"
 #include "game/shell.h"
 #include "log.h"
+
+#define M_INITIAL_FOV 65
 
 #define L_DEFAULT_VIEWPORT                                                     \
     { .width = SHELL_HEADLESS_WIDTH, .height = SHELL_HEADLESS_HEIGHT }
@@ -14,7 +18,75 @@ static VIEWPORT_RECT m_Rects[VIEWPORT_NUMBER_OF] = {
 };
 #undef L_DEFAULT_VIEWPORT
 
-VIEWPORT_RECT *g_Viewport_Rects = m_Rects;
+static int16_t m_CurrentFOV = M_INITIAL_FOV;
+
+void Viewport_Init(int32_t x, int32_t y, int32_t width, int32_t height)
+{
+    const VIEWPORT_RECT *const target = &m_Rects[VIEWPORT_TARGET];
+    VIEWPORT_RECT *const game = &m_Rects[VIEWPORT_GAME];
+    VIEWPORT_RECT *const ui = &m_Rects[VIEWPORT_UI];
+
+    if (x < 0 || y < 0 || width < 0 || height < 0) {
+        struct {
+            int32_t w, h;
+        } ar;
+        switch (g_Config.rendering.aspect_mode) {
+        case AM_4_3:
+            ar.w = 4;
+            ar.h = 3;
+            break;
+        case AM_16_9:
+            ar.w = 16;
+            ar.h = 9;
+            break;
+        case AM_16_10:
+            ar.w = 16;
+            ar.h = 10;
+            break;
+        case AM_ANY:
+            ar.w = target->width;
+            ar.h = target->height;
+            break;
+        }
+
+        x = 0;
+        y = 0;
+        width = target->width;
+        height = target->height;
+        if (g_Config.rendering.aspect_mode != AM_ANY) {
+            width = height * ar.w / ar.h;
+        }
+    }
+
+    ui->x = x;
+    ui->y = y;
+    ui->width = width;
+    ui->height = height;
+
+    game->x = x;
+    game->y = y;
+    game->width = width / g_Config.rendering.upscaling_factor;
+    game->height = height / g_Config.rendering.upscaling_factor;
+
+    g_PhdLeft = Viewport_GetMinX(VIEWPORT_GAME);
+    g_PhdTop = Viewport_GetMinY(VIEWPORT_GAME);
+    g_PhdRight = Viewport_GetMaxX(VIEWPORT_GAME);
+    g_PhdBottom = Viewport_GetMaxY(VIEWPORT_GAME);
+}
+
+int16_t Viewport_GetSystemFOV(void)
+{
+    return m_CurrentFOV;
+}
+
+int16_t Viewport_GetUserFOV(void)
+{
+#if TR_VERSION == 1
+    return g_Config.visuals.fov_value * DEG_1;
+#elif TR_VERSION == 2
+    return g_Config.visuals.fov * DEG_1;
+#endif
+}
 
 int16_t Viewport_GetEffectiveFOV(void)
 {
@@ -67,7 +139,7 @@ VIEWPORT_RECT Viewport_GetRect(const VIEWPORT_SPACE space)
     return m_Rects[space];
 }
 
-void Viewport_ResetCommon(void)
+void Viewport_Reset(void)
 {
     const SHELL_SIZE size = Shell_GetCurrentSize();
     VIEWPORT_RECT *const window = &m_Rects[VIEWPORT_WINDOW];
@@ -113,6 +185,13 @@ void Viewport_ResetCommon(void)
     }
     target->x = (window->width - target->width) / 2;
     target->y = (window->height - target->height) / 2;
+    Viewport_Init(-1, -1, -1, -1);
+    Viewport_Debug();
+}
+
+void Viewport_AlterFOV(int16_t fov)
+{
+    m_CurrentFOV = fov;
 }
 
 void Viewport_Debug(void)
