+++ src/libtrx/include/libtrx/gfx/context.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "2d/2d_renderer.h"
-#include "3d/3d_renderer.h"
-#include "common.h"
-#include "game/viewport.h"
-#include "renderer.h"
+#include "../game/viewport.h"
+#include "./2d/2d_renderer.h"
+#include "./3d/3d_renderer.h"
+#include "./common.h"
+#include "./renderer.h"
 
 #include <stdint.h>
