+++ src/tr1/game/console/common.c
@@ -1,10 +1,9 @@
 #include "game/console/common.h"
 
-#include "game/viewport.h"
-
 #include <libtrx/game/output.h>
 #include <libtrx/game/scaler.h>
 #include <libtrx/game/ui.h>
+#include <libtrx/game/viewport.h>
 
 void Console_DrawBackdrop(void)
 {
