+++ src/tr1/game/level.c
@@ -11,7 +11,6 @@
 #include "game/shell.h"
 #include "game/sound.h"
 #include "game/stats.h"
-#include "game/viewport.h"
 #include "global/types.h"
 #include "global/vars.h"
 
@@ -29,6 +28,7 @@
 #include <libtrx/game/option.h>
 #include <libtrx/game/output.h>
 #include <libtrx/game/overlay.h>
+#include <libtrx/game/viewport.h>
 #include <libtrx/log.h>
 #include <libtrx/memory.h>
 #include <libtrx/utils.h>
