+++ src/tr1/meson.build
@@ -224,7 +224,6 @@ sources = [
   'game/stats/common.c',
   'game/ui/dialogs/stats.c',
   'game/ui/elements/bar.c',
-  'game/viewport.c',
   'global/enum_map.c',
   'global/vars.c',
   resources,
