+++ src/tr2/decomp/decomp.c
@@ -5,7 +5,6 @@
 #include "game/lara.h"
 #include "game/level.h"
 #include "game/objects/vars.h"
-#include "game/viewport.h"
 #include "global/vars.h"
 
 #include <libtrx/config.h>
@@ -15,6 +14,7 @@
 #include <libtrx/game/game_string_table.h>
 #include <libtrx/game/music.h>
 #include <libtrx/game/output.h>
+#include <libtrx/game/viewport.h>
 #include <libtrx/utils.h>
 
 // TODO: delegate these constants to individual vehicle code
