+++ src/tr2/game/shell/common.c
@@ -3,7 +3,6 @@
 #include "game/level.h"
 #include "game/savegame.h"
 #include "game/sound.h"
-#include "game/viewport.h"
 
 #include <libtrx/config.h>
 #include <libtrx/debug.h>
@@ -13,6 +12,7 @@
 #include <libtrx/game/output.h>
 #include <libtrx/game/overlay.h>
 #include <libtrx/game/shell.h>
+#include <libtrx/game/viewport.h>
 #include <libtrx/gfx/context.h>
 #include <libtrx/memory.h>
 #include <libtrx/strings.h>
