--- src/tr2/game/viewport.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/viewport.h>
-
-void Viewport_Init(int32_t x, int32_t y, int32_t width, int32_t height);
